chore: improved readme accessibility and usability
Implemented a summary in the repository readme to ease getting information and improve the readability. It will also be useful to improve the accessibility of screen readers, since in addition to the table of contents, options to return to the table of contents have been implemented at important points where information is finalized.